// src/views/welcome/utils/types.ts
// 虽然字段很少 但是抽离出来 后续有扩展字段需求就很方便了

interface FormItemProps {
  _id: string;
  /** 角色名称 */
  name: string;
  /** 角色编号 */
  code: string;
  /** 备注 */
  remark: string;

  imageUrl: string;
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };